Implement automatic configuration flow
This update implements an automated configuration generation flow in the start-up script. It processes variables prefixed with 'PDNS_' and accordingly creates or adjusts the configuration. Additionally, GitHub workflows for auto-release, build-testing, and Docker publishing have been established. A variety of .gitignore files and necessary configurations have been added as well.